+++ R/KHfilgruppe.R
@@ -12,6 +12,7 @@ LagFilgruppe <- function(name, write = TRUE, dumps = list(), qualcontrol = TRUE)
   user_args = as.list(environment())
   parameters <- get_filegroup_parameters(user_args = user_args)
   parameters[["old_locale"]] <- ensure_utf8_encoding()
+  parameters[["threads"]] <- set_threads()
   # For dev and debug: use SetFilgruppeParameters("NAME") and run step by step below
   if(parameters$write) sink(file = file.path(getOption("khfunctions.root"), getOption("khfunctions.fgdir"), getOption("khfunctions.fg.logg"), paste0(parameters$name, "_", parameters$batchdate, "_LOGG.txt")), split = TRUE)
   if(parameters$n_files == 0) stop("Ingen originalfiler funnet, filgruppe kan ikke genereres. Sjekk at staving matcher for alle relevante felter i ACCESS")
@@ -58,6 +59,10 @@ lagfilgruppe_cleanup <- function(parameters){
   if(parameters$old_locale != "nb-NO.UTF-8") Sys.setlocale("LC_ALL", parameters$old_locale)
   RODBC::odbcCloseAll()
   if(exists("org_geo_codes", envir = .GlobalEnv)) rm(org_geo_codes, envir = .GlobalEnv)
+  if(!is.null(parameters$threads)){
+    data.table::setDTthreads(parameters$threads$dt)
+    collapse::set_collapse(nthreads = parameters$threads$collapse)
+  }
 }
 
 #' @title initiate_cleanlog