+++ NEWS.md
@@ -10,12 +10,13 @@
 - Fixed error message when original file not found ([issue151](https://github.com/helseprofil/khfunctions/issues/151), @raniets)
 - Implemented function `control_rate_lks` ([issue128](https://github.com/helseprofil/khfunctions/issues/128))
 - `fix_geo_special` no longer delete values. 
+- `do_special_handling` gains dt_name argument, to ensure correct name in code environment. 
+- Implemented new function `set_threads`. In `LagKUBE`/`LagFilgruppe`, data.table and collapse now utilizes more threads to speed up processing. 
 
 ## Code cleaning
 - Remove get(col)/mget(col), replace with dt[[col]] or .SDcols, whichever is most relevant
 - Implemented data.table::set() several places replacing := or let()
 - More functions operating by reference, avoiding unneccessary copies
-- Explicitly provide dt_name to do_special_handling. 
 
 # khfunctions 1.1.6 (2026-02-25)